+++ core/src/testFixtures/kotlin/com/example/typespec/TypeSpecBasePlatformTestCase.kt
@@ -1,0 +1,16 @@
+package com.example.typespec
+
+import com.intellij.testFramework.fixtures.BasePlatformTestCase
+
+abstract class TypeSpecBasePlatformTestCase : BasePlatformTestCase() {
+    override fun setUp() {
+        suppressUltimatePostStartupConstructorErrors()
+        super.setUp()
+    }
+
+    companion object {
+        init {
+            suppressUltimatePostStartupConstructorErrors()
+        }
+    }
+}